Improvement

+++ routes/Main.js
@@ -27,7 +27,7 @@ import { socket } from "../socket";
 import { getUser } from "../api";
 import { readFile, removeFile, writeFile } from "../helpers/offline";
 import { changeDate } from "../helpers/libs";
-import cleanData from '../helpers/cleanData';
+import cleanData from "../helpers/cleanData";
 import * as BackgroundFetch from "expo-background-fetch";
 import * as TaskManager from "expo-task-manager";
 
@@ -164,10 +164,15 @@ const Main = () => {
       });
 
       if (res.error) {
-        navigation.current.replace("SignIn");
-        await removeFile({ name: 'data.json' });
-        await removeFile({ name: 'user.json' });
-        return cleanData(dispatch);
+        if (!activeGroup.active) {
+          navigation.current.replace("SignIn");
+          await removeFile({ name: "data.json" });
+          await removeFile({ name: "user.json" });
+          return cleanData(dispatch);
+        } else {
+          const information = await readFile({ name: "user.json" });
+          changeGeneralInformation(dispatch, information);
+        }
       }
 
       const data = await readFile({ name: "data.json" });
